Moved wrappers to their own file, added ImplicitPrognosticWrapper
