+++ sympl/_core/util.py
@@ -1,11 +1,12 @@
+from datetime import datetime
+
+import numpy as np
+from six import string_types
+
+from .array import DataArray
 from .exceptions import (
     SharedKeyError, InvalidStateError, InvalidPropertyDictError)
-from .constants import default_constants
-from .array import DataArray
-from six import string_types
-import numpy as np
-from datetime import datetime
-import warnings
+
 try:
     from numba import jit
 except ImportError:
@@ -205,7 +206,7 @@ def get_numpy_array(
         Length 1 dimensions will be created if the dimension is 'x', 'y', 'z',
         or '*' and does not exist in data_array. 'x', 'y', and 'z' indicate any axes
         registered to those directions with
-        :py:function:`~sympl.set_dimension_names`. '*' indicates an axis
+        :py:function:`~sympl.set_direction_names`. '*' indicates an axis
         which is the flattened collection of all dimensions not explicitly
         listed in out_dims, including any dimensions with unknown direction.
     return_wildcard_matches : bool, optional
@@ -360,10 +361,12 @@ def restore_data_arrays_with_properties(
                 result_attrs=attrs)
         except ShapeMismatchError:
             raise InvalidPropertyDictError(
-                'quantity {} has dims_like {}, but the provided array for {} has '
-                'a shape incompatible with the shape of {} in the input '
-                'quantities. Do they really have the same dimensions?'.format(
-                    quantity_name, dims_like, quantity_name, dims_like
+                'output quantity {} has dims_like input {}, but the '
+                'provided output array for {} has '
+                'a shape {} incompatible with the input shape {} of {}. '
+                'Do they really have the same dimensions?'.format(
+                    quantity_name, dims_like, quantity_name, array.shape,
+                    result_like.shape, dims_like
                 )
             )
     return out_dict
@@ -384,7 +387,7 @@ def restore_dimensions(array, from_dims, result_like, result_attrs=None):
         argument used in the call to get_numpy_array.
         'x', 'y', and 'z' indicate any axes
         registered to those directions with
-        :py:function:`~sympl.set_dimension_names`. '*' indicates an axis
+        :py:function:`~sympl.set_direction_names`. '*' indicates an axis
         which is the flattened collection of all dimensions not explicitly
         listed in out_dims, including any dimensions with unknown direction.
     result_like : DataArray
@@ -443,13 +446,6 @@ def same_list(list1, list2):
     items present in list2 (ignoring order)."""
     return (len(list1) == len(list2) and all(
         [item in list2 for item in list1] + [item in list1 for item in list2]))
-
-
-def set_dimension_names(x=None, y=None, z=None):
-    warnings.warn(
-        'set_dimension_names is deprecated, use set_direction_names instead',
-        DeprecationWarning)
-    return set_direction_names(x=x, y=y, z=z)
 
 
 def set_direction_names(x=None, y=None, z=None):
@@ -528,125 +524,6 @@ def _ensure_no_invalid_directions(out_dims):
     if len(invalid_dims) != 0:
         raise ValueError(
             'Invalid direction(s) in out_dims: {}'.format(invalid_dims))
-
-
-class UpdateFrequencyWrapper(object):
-    """
-    Wraps a prognostic object so that when it is called, it only computes new
-    output if sufficient time has passed, and otherwise returns its last
-    computed output. The Delayed object requires that the 'time' attribute is
-    set in the state, in addition to any requirements of the Prognostic
-
-    Example
-    -------
-    This how the wrapper should be used on a fictional Prognostic class
-    called MyPrognostic.
-
-    >>> from datetime import timedelta
-    >>> prognostic = UpdateFrequencyWrapper(MyPrognostic(), timedelta(hours=1))
-    """
-
-    def __init__(self, prognostic, update_timedelta):
-        """
-        Initialize the UpdateFrequencyWrapper object.
-
-        Args
-        ----
-        prognostic : Prognostic
-            The object to be wrapped.
-        update_timedelta : timedelta
-            The amount that state['time'] must differ from when output
-            was cached before new output is computed.
-        """
-        self._prognostic = prognostic
-        self._update_timedelta = update_timedelta
-        self._cached_output = None
-        self._last_update_time = None
-
-    def __call__(self, state, **kwargs):
-        if ((self._last_update_time is None) or
-                (state['time'] >= self._last_update_time +
-                 self._update_timedelta)):
-            self._cached_output = self._prognostic(state, **kwargs)
-            self._last_update_time = state['time']
-        return self._cached_output
-
-    def __getattr__(self, item):
-        return getattr(self._prognostic, item)
-
-
-class TendencyInDiagnosticsWrapper(object):
-    """
-    Wraps a prognostic object so that when it is called, it returns all
-    tendencies in its diagnostics.
-
-    Example
-    -------
-    This how the wrapper should be used on a fictional Prognostic class
-    called RRTMRadiation.
-
-    >>> prognostic = TendencyInDiagnosticsWrapper(RRTMRadiation(), 'radiation')
-    """
-
-    def __init__(self, prognostic, label):
-        """
-        Initialize the Delayed object.
-
-        Args
-        ----
-        prognostic : Prognostic
-            The object to be wrapped
-        label : str
-            Label describing what the tendencies are due to, to be
-            put in the diagnostic quantity names.
-        """
-        self._prognostic = prognostic
-        self._tendency_label = label
-        self._tendency_diagnostic_properties = {}
-        for quantity_name, properties in prognostic.tendency_properties.items():
-            diagnostic_name = 'tendency_of_{}_due_to_{}'.format(quantity_name, label)
-            self._tendency_diagnostic_properties[diagnostic_name] = properties
-
-    @property
-    def inputs(self):
-        return list(self.diagnostic_properties.keys())
-
-    @property
-    def input_properties(self):
-        return_dict = self._prognostic.input_properties.copy()
-        return_dict.update(self._tendency_diagnostic_properties)
-        return return_dict
-
-    @property
-    def diagnostics(self):
-        return list(self.diagnostic_properties.keys())
-
-    @property
-    def diagnostic_properties(self):
-        return_dict = self._prognostic.diagnostic_properties.copy()
-        return_dict.update(self._tendency_diagnostic_properties)
-        return return_dict
-
-    def __call__(self, state, **kwargs):
-        tendencies, diagnostics = self._prognostic(state, **kwargs)
-        for quantity_name in tendencies.keys():
-            diagnostic_name = 'tendency_of_{}_due_to_{}'.format(
-                quantity_name, self._tendency_label)
-            diagnostics[diagnostic_name] = tendencies[quantity_name]
-        return tendencies, diagnostics
-
-    def __getattr__(self, item):
-        return getattr(self._prognostic, item)
-
-
-def replace_none_with_default(constant_name, value):
-    """If value is None, returns the default constant for the constant name.
-    Otherwise, returns value. If the default constant is not defined, raises
-    KeyError."""
-    if value is None:
-        return default_constants[constant_name]
-    else:
-        return value
 
 
 def update_dict_by_adding_another(dict1, dict2):
